% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{agglom_tree}
\alias{agglom_tree}
\title{Agglomerative decision tree}
\usage{
agglom_tree(
  data,
  original_data = original_data[training_indices, ],
<<<<<<< HEAD
  training_set,
=======
>>>>>>> afb80ee0935fb3ec0dd7caacaef56f1ca436547f
  re_use_variables = F
)
}
\arguments{
\item{original_data}{Data before the dummyset splits}
}
\description{
A (novel?) agglomerative decision tree. It is built bottom-up, with one final leaf per model. Used for interpreting ICR results.
}
