% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{autotune}
\alias{autotune}
\title{Autotune}
\usage{
autotune(
  input_data,
  dv,
  iv,
  restrict_sample = T,
  random_seed,
  training_indices = NULL
)
}
\arguments{
\item{input_data}{Input the same input data set as you will use for the main icr analysis.}
}
\description{
Generate automatically a set of tuning parameters for a particular modeling problem.
}
\examples{
 test <- benchmarking(random_seeds=100001:100025, n_cases=10000)
}
\keyword{Autotune}
\keyword{optimization}
\keyword{parameters.}
