% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{benchmarking}
\alias{benchmarking}
\title{Benchmarking}
\usage{
benchmarking(
  random_seed = c(123),
  n_random_iterations,
  generations = c(40, 200, 1000),
  solution_thresh = c(0.01),
  n_strands = c(50),
  real_models = c(20),
  error_sd = c(0.05),
  n_IVs = c(30),
  n_cases = 5000,
  mutation_rate = 0.2,
  n_children = 2,
  test_set_model_closeness = c("residuals"),
  parallelize = F,
  detailed_plots_on = F,
  boxplots = T,
  starting_population,
  hold_back_cores = 2,
  death_by_ageing = 60,
  nelitism = 10,
  target_training_percent = NULL,
  generation_limit = NULL,
  final_model_aggregation = T
)
}
\arguments{
\item{generations}{Benchmarking parameter. A vector of integers specifying the number of generations for initial ICR genetic training.}

\item{solution_thresh}{Benchmarking parameter. A vector of positive numeric values specifying the discrete solution target on each side of the solution, where solutions may range from 0 to 1.}

\item{n_strands}{Benchmarking parameter. A vector of integers indicating the number of population that survive each generation of the genetic training algorithm.}

\item{real_models}{Benchmarking parameter. A vector of integers specifying the 'ground truth' number of underlying models in synthetic data.}

\item{error_sd}{Benchmarking parameter. A vector of numeric values indicating the standard deviation of error of synthetic data. Synthetic data range from 0 to 1.}

\item{n_IVs}{Benchmarking parameter. A vector of integers specifying the 'ground truth' number of underlying relevant independent variables for the given dependent variable in synthetic data.}

\item{n_cases}{Integer value. The number of overall cases in the synthetic data.}

\item{test_set_model_closeness}{Do not change. Will be removed as a parameter.}

\item{parallelize}{T indicates that the benchmarking will happen via multiple cores (sockets), whereas F indicates sequential.}

\item{detailed_plots_on}{Produces bi-dimensional X-Y plots for various steps of the ICR algorithm. Only useful for low-dimensional analysis.}

\item{boxplots}{Produces boxplots comparing method outcomes with one another at the end of the benchmarking routine.}

\item{starting_population}{An integer representing the number of starting population.}

\item{hold_back_cores}{Prevents X cores from being involved in the parallel processing, thus reserving computing power.}

\item{random_seeds}{Benchmarking parameter. A vector of integers to serve as random seeds.}
}
\description{
Run a random set of generated (fake) data and comparing the results across multiple methods
}
\examples{
 test <- benchmarking(random_seeds=100001:100025, n_cases=10000)
}
\keyword{Benchmarking}
\keyword{Comparison}
