% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{biggest_anova_gap}
\alias{biggest_anova_gap}
\title{Biggest anova gap}
\usage{
biggest_anova_gap(ivs, these_models, populations, F_threshold = 1.25)
}
\description{
Identifies largest ANOVA gap between subgroups. Part of agglomerative tree building process.
}
