% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{check_profile_performance}
\alias{check_profile_performance}
\title{Deprecated?}
\usage{
check_profile_performance(original_profiles, comparison_profiles)
}
\description{
Track how well the profiles are being correctly found or not
}
