% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{combine_leafs}
\alias{combine_leafs}
\title{Combine leaves}
\usage{
combine_leafs(split_variable)
}
\description{
This function combines leafs based on a particular variable split. In an agglomerative manner this builds a population-based tree from the bottom up. Part of the agglomerative tree building proces
}
