% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{compare_data}
\alias{compare_data}
\title{Compare data}
\usage{
compare_data(
  true_sample = sample.data$profile,
  found_sample = true_regression_case_based$dna_profiles,
  true_case_assignments = sample.data$data$model,
  found_case_assignments = true_regression_case_based$model_assignment,
  dna_pool = sample.data$models,
  found_models = true_regression_case_based$dna_pool
)
}
\description{
Compare original sample characteristics and original model characteristics with outcome####
}
