% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{condense_models}
\alias{condense_models}
\title{Condenses models.}
\usage{
condense_models(
  data = NULL,
  training_set = training_set,
  method = c("best", "cluster", "genetic"),
  nmodel = 10,
  auto_cluster = TRUE,
  closeness_threshold = solution_threshold,
  solution_table,
  no_correlation_thresh = -0.1,
  pearson_corr_threshold = 0.7
)
}
\description{
Condenses models.
}
