% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{create_profile}
\alias{create_profile}
\title{Create Profile}
\usage{
create_profile(
  data_after_dummies = df,
  model_assignment_data = NULL,
  dummy_vars = "cntry",
  model_labels = NULL
)
}
\description{
Create a descriptive profile of the subpopulations for each model, including all categories (reference categories are not left out)
}
