% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ess2016}
\alias{ess2016}
\title{Data from the 2016 European Social Survey}
\format{
\subsection{\code{ess2016}}{

A data frame with 44387 rows 8 columns:
\describe{
\item{happy}{A higher value means greater happiness: "Taking all things together, how happy would you say you are?"}
\item{female}{Dichotomous, a higher value indicates female.}
\item{age}{Age in years.}
\item{cntry}{Country.}
\item{ppltrst}{a higher value indicates higher generalized trust: "enerally speaking, would you say that most people can be trusted, or that you can’t be too careful in dealing with people? }
\item{sclmeet}{A higher value indicates higher frequency of social meetings: "how often do you meet socially with friends, relatives or work colleagues?" (Never, Less than once a month, Once a month, Several times a month, Once a week, Several times a week, Every day)}
\item{income}{Income in deciles.}
\item{health}{Self-rated health: "How is your health37 in general?"}
...
}
}
}
\source{
European Social Survey European Research Infrastructure (ESS ERIC). (2020). ESS8 - integrated file, edition 2.2 \link{Data set}. Sikt - Norwegian Agency for Shared Services in Education and Research. https://doi.org/10.21338/ESS8E02_2
}
\usage{
ess2016
}
\description{
An extract of cross-sectional data from the 2016 European Social Survey data is used in order to demonstrate the use of the ICR Method for educational purposes.
Full documentation of the ESS and its data are available on its website: https://ess-search.nsd.no/en/study/f8e11f55-0c14-4ab3-abde-96d3f14d3c76.
Included variables are a limited selection of those with a theoretical and empirical relationship to happiness.
}
\keyword{datasets}
