% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{interpret_by_case}
\alias{interpret_by_case}
\title{Used by case-based decision tree.}
\usage{
interpret_by_case(
  data,
  models,
  profiles,
  categorical_variables = NULL,
  reference_categories = NULL,
  min_bucket = 0,
  max_depth = NULL,
  eliminate_multileaves = T,
  assignments = NULL
)
}
\description{
Used by case-based decision tree.
}
