% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{interpret_results}
\alias{interpret_results}
\title{Deprecated?}
\usage{
interpret_results(
  dna = dna_pool,
  profiles = dna_profiles,
  solution_table = solution_table
)
}
\description{
out of date.. Potentially use the proportional explanation table as a useful statistic
}
