% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{model_closeness}
\alias{model_closeness}
\title{Model closeness}
\usage{
model_closeness(
  data = training_set,
  profiles = dna_profiles,
  models = models,
  method = NULL,
  cutlist = cutlist,
<<<<<<< HEAD
  population_tree_results = NULL,
  training_set = training_set,
  testing_set = testing_set,
  training_indices = training_indices
=======
  population_tree_results = NULL
>>>>>>> afb80ee0935fb3ec0dd7caacaef56f1ca436547f
)
}
\description{
Assigns models to cases. (e.g. for predictions for new cases)
}
