% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{model_plot}
\alias{model_plot}
\title{Model plot}
\usage{
model_plot(
  data_with_models,
  label,
  base_line_plot = F,
  models,
  turned_on = F,
  X_variable = 1,
  point_labels = T,
  local_regression_line = F
)
}
\description{
Plot group-based bivariate regression lines
}
