% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{predict_new}
\alias{predict_new}
\title{Predict new}
\usage{
predict_new(
  models = result_training$dna_pool,
  profiles = result_training$dna_profiles,
  new_data = testing_set,
  closeness_threshold = solution_threshold,
  assign_models = T,
  method = "case-based",
  population_tree_results = NULL,
  dv = T,
  training_set = training_set,
  testing_set = testing_set,
  training_indices = training_indices
)
}
\description{
Use ICR results to predict never-seen case dependent variables
}
