% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{sample_data}
\alias{sample_data}
\title{Create Sample Data}
\usage{
sample_data(
  c.number = 300,
  m.number = 5,
  m_sizes = c(100, 50, 50, 50, 50),
  bin.v.number = 5,
  con.v.number = 5,
  error = TRUE,
  error.sd = 1,
  data.form = c("raw", "standardized", "normalized"),
  messy_subgroups = T,
  min_variable_split = 0.25
)
}
\description{
Produces synthetic data
}
