% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{test_children}
\alias{test_children}
\title{Genetic Algorithm Training}
\usage{
test_children(
  dna_pool = children,
  input_data = training_set,
  closeness_threshold = solution_threshold,
  generations = 10,
  difficulty = 0.8,
  death_rating_thresh = 0.5,
  death_number_thresh = NULL,
  uniqueness_power = 2,
  number_of_children = n_children,
  death_by_ageing = 999999999,
  min_strand_solution_proportion = 0.05,
  mutations = mutation_rate,
  mating = c("pam", "dam", "random"),
  mating_power = 1,
  selection = c("threshold", "probability"),
  rem_doppleganger = T,
  nelitism = 0
)
}
\description{
Run genetic algorithm and store average value characteristics for children###
This first checks the exact algo for a solution. Then loops through children to see if they can find it. Then it ranks the children according to the uniqueness of their solution.
}
