% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{transform_dummies}
\alias{transform_dummies}
\title{Transform dummies}
\usage{
transform_dummies(
  x,
  reference = NULL,
  group_terms = NULL,
  reference_cat = NULL
)
}
\description{
Transforms dummysets back to standard variables
}
