% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{combinatorial_columns}
\alias{combinatorial_columns}
\title{Create combinatorial columns}
\usage{
combinatorial_columns(input_data)
}
\arguments{
\item{input_data}{Dataframe to be converted into combinatorial columns.}
}
\description{
Creates combinatorial columns for use in the agglomerative decision tree
}
