% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{create_profile}
\alias{create_profile}
\title{Create Profile}
\usage{
create_profile(
<<<<<<< HEAD
  data_after_dummies = original_data,
=======
  data_after_dummies = df,
>>>>>>> afb80ee0935fb3ec0dd7caacaef56f1ca436547f
  model_assignment_data = NULL,
  dummy_vars = "cntry",
  model_labels = NULL
)
}
\description{
Create a descriptive profile of the subpopulations for each model, including all categories (reference categories are not left out)
}
