% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRegress.R
\name{icr}
\alias{icr}
\title{ICR}
\usage{
icr(
  generations = c(1000),
  solution_thresh = c(0.01),
  n_strands = c(50),
  detailed_plots_on = F,
  input_data,
  dv,
  iv,
  starting_pop = 10000,
  mutation_rate = 0.2,
  n_children = 2,
  death_by_ageing = 60,
  nelitism = 10,
  force_subgroup_n = NA,
  re_use_variables = F,
  .tuning = F,
  autotune_params = NULL,
  max_n_condensations = 6,
  force_n_condensation = NULL,
  training_indices = NULL,
  target_training_percent = NULL,
  generation_limit = NULL,
<<<<<<< HEAD
  final_model_aggregation = T,
  training_set = NULL,
  testing_set = NULL
=======
  final_model_aggregation = T
>>>>>>> afb80ee0935fb3ec0dd7caacaef56f1ca436547f
)
}
\arguments{
\item{generations}{An integer specifying the number of generations for initial ICR genetic training.}

\item{solution_thresh}{A positive numeric value specifying the discrete solution target on each side of the solution, where solutions may range from 0 to 1.}

\item{n_strands}{Survivors. An integer indicating the number of population that survive each generation of the genetic training algorithm.}

\item{detailed_plots_on}{Produces bi-dimentional X-Y plots for various steps of the ICR algorithm. Only useful for low-dimensional analysis.}

\item{input_data}{A pure data.frame containing the dependent and independent variables. If not specified otherwise, the first column will be used as the dependent variable. If this df contains a column called 'model', it will be removed. All dummy variables must be coded as numerical and 0-1, factors of more than 3 categories should be entered as char columns.}

\item{dv}{The dependent variable. If not specified, the first column is used.}

\item{starting_pop}{Starting sets of random coefficients (population, children) in first generation of genetic algorithm.}

\item{force_subgroup_n}{The number of model-relevant subgroups may be manually set with an integer, thus ignoring the automatic identification of of the ideal number of subgroups. The composition of chosen n subgroups will still be automatically determined.}

\item{re_use_variables}{Allows the same variable to be split multiple times if possible and relevant during the agglomeration tree step.}

\item{autotune_params}{A list object exported from autotune(). If an autotune object is provided, it overrides the settings of solution_thresh, n_strands, starting_pop, mutation_rate, n_children, and nelitism.}

\item{ivs}{The independent variable. If not specified, all variables besides the dv are used.}
}
\description{
Run the ICR algo
}
\examples{
 icr_results <- icr(generations=400,
input_data=ess2016,
dv="happy",
iv=c("female", "age", "cntry", "ppltrst", "sclmeet",      "income", "health"),
starting_pop=10000, #1000
n_strands = 100,
mutation_rate = .02,
solution_thresh = .003,
n_children=3,
death_by_ageing =60,
nelitism=10)
}
\keyword{ICR}
